Test how dataframes are the same / different.

df_diff prints out how dataframes are the same and how they are different.

df_diff(x, y)

Arguments

x

First dataframe.

y

Second dataframe.

Value

Test results and differences are printed to standard out.

Details

Tests include

Equal # of rows

Equal # of columns

Row names are the same

Column names are the same

Class(es) is / are the same

Column types are the same - work in progress

Values within each column are the same

df_diff is unstable at the moment and should be considered a work in progress.

Examples

x <- tibble::tribble(
 ~col1, ~col2,
 1, "a",
 2, "b",
 3, "c"
)

y <- data.frame(col1 = 1:3, col2 = c("a", "b", "d"), col3 = c(10, 11, 12))

df_diff(x, y)

#> The number of rows is equal
#> The number of columns is not equal
#>    x has 2 columns
#>    y has 3 columns
#> The row names are the same
#> The column names are not the same
#>    x has column names: col1 col2 
#>    y has column names: col1 col2 col3 
#> The class(es) is / are not the same
#>    x has class(es): tbl_df tbl data.frame 
#>    y has class(es): data.frame 
#> The number of columns is not equal; testing will occur on columns that are named the same
#>   column: col1 
#>    all equal: Cols in y but not x: 1:3. Cols in x but not y: col1.  
#>   column: col2 
#>    all equal: Cols in y but not x: structure(1:3, .Label = c("a", "b", "d"), class = "factor"). Cols in x but not y: col2.  
#>