List exact names of columns.

exact_cols returns the exact column names, including the case.

exact_cols(tbl, cols)

Arguments

tbl

A tibble or dataframe.

cols

A character vector of column names.

Value

A character vector of column names with the appropriate case.

Details

This particularly useful when working with Microsoft SQL Server tables as the column names may be case sensitive. This is also useful when working with table metadata and the exact column names are required for matching purposes.

tbl_vars does not return case sensitive columns, thus the the need for this function.

Examples

## create tibble
my_tibble <- tibble::tribble(
  ~MixedCase, ~lowercase, ~UPPERCASE, ~aLtCaSe,
  1, 11, 111, 1111,
  2, 22, 222, 2222,
  3, 33, 333, 3333
)

## column names to search for, ignoring case
my_cols <- c("mixedcase", "altcase")

my_cols_case <- my_tibble %>%
  exact_cols(my_cols)

## particularly useful when combined with dplyr::one_of
my_tibble %>%
  dplyr::select(dplyr::one_of(my_cols_case))

#> # A tibble: 3 × 2
#>   MixedCase aLtCaSe
#>       <dbl>   <dbl>
#> 1         1    1111
#> 2         2    2222
#> 3         3    3333
#>