Generate SQL query.

sql_list_sql_ returns an SQL query as a string based on the parameters passed to it.

sql_list_sql_(db_con, db_name, db_schema, obj)

Arguments

db_con

Database server connection object.

db_name

Database name.

db_schema

Database schema.

obj

String used with switch to select appropriate SQL query. Choices include

"schemas"

"tables"

"views"

"tvs"

Value

An SQL query as a string.

Details

sql_list_sql_ is used along with the function factory sql_list_ to create functions that return properties of a Microsoft SQL Server database.

The intent is not to use this function directly. Instead use one of the functions

sql_list_schemas

sql_list_tbls

sql_list_views

sql_list_tvs