Get XML Text.

Return an XML text value if provided the document and Xpath.

xml_get_text(x, xpath)

Arguments

x

XML document: a literal XML document, a URL, or a string.

xpath

A string containing a xpath (1.0) expression.

Value

XML text as a string or NA_character_ if an error occurs.

Details

Serves as a convenience function. Utilizes xml_find_first and xml_text and will return NA_character_ if the XML text cannot be located or an error occurs.

Because the function uses xml_find_first, only the first match is returned.

Examples

xml_doc <- "
<root_tag>
  <another_tag>
    <var_text>10</var_text>
    <var_attr value=\"10\"></var_attr>
  </another_tag>
</root_tag>
"

ten <- xml_get_text(xml_doc,
                    "//root_tag/another_tag/var_text")
identical(ten, "10")

#> [1] TRUE
#>